refactor(features): moved features up
