+++ front/features/Login/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Button } from '../../ui/Button';
+import { Button } from '../../components/ui/Button';
 
 interface Props {}
